Keep panel lifecycle cleanup outside the Workspace core
CodeRabbit flagged the panel teardown and agent runtime block as too large for Workspace.swift. The behavior stays the same, but the new lifecycle subsystem now lives in a focused Workspace extension so the PR no longer expands the already-large core file by the full helper block.

Constraint: CodeRabbit pre-merge boundary check required extracting Workspace+PanelLifecycle.swift

Rejected: Leave the helper block in Workspace.swift | kept the PR blocked on file-boundary review

Confidence: high

Scope-risk: narrow

Tested: ./scripts/reload.sh --tag issue-3741-coderabbit

Not-tested: Local unit tests per repository policy
